backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.10.0
openai==1.54.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
jsonschema==4.23.0
tiktoken==0.8.0
datasets==3.2.0
huggingface_hub==0.27.0
python-multipart==0.0.20

